*** Se modifico el componente de ListarEmpleados en el proyecto React JS
Se corrigio los texto en botones, estilos css, acciones.

+++ react-app/src/components/ListarEmpleados.jsx
@@ -30,12 +30,18 @@ class ListarEmpleados extends Component{
     }
 
     fncEliminarEmpleado(empleadoId){
+
+        const re = window.confirm("¿Seguro que deseas eliminar al empleado seleccionado?");
+        if(!re) return null;
+
         EmpleadoService.deleteEmpleado(empleadoId)
         .then((resp) => {
+            alert("Empleado eliminado exitosamente.");
             window.location.href = "/empleados";
         }).catch((resp) => {
             window.location.href = "/empleados";
         });
+
     }
 
     fncRegistrarEmpleado(){
